feat(cat-data-service): Refactor ready to call axum directly when poem returns 404
