+++ src/cat-data-service/Cargo.toml
@@ -45,21 +45,12 @@ poem-extensions = { version = "0.7.2" }
 # Metrics - Poem
 opentelemetry-prometheus = { version = "0.13" }
 opentelemetry = { workspace = true }
-
-once_cell = { workspace = true }
-
 cryptoxide = { workspace = true }
-
 uuid = { workspace = true }
-
 lazy_static = { workspace = true }
-
 url = { workspace = true }
-
 dotenvy = { workspace = true }
-
 panic-message = "0.3"
-
 regex = { workspace = true }
 
 [dev-dependencies]
