+++ src/cat-data-service/src/axum/v0/fund.rs
@@ -1,8 +1,4 @@
-use crate::{
-    service::{handle_result, Error},
-    state::State,
-    types::SerdeType,
-};
+use crate::{axum::handle_result, service::Error, state::State, types::SerdeType};
 use axum::{routing::get, Router};
 use event_db::types::vit_ss::fund::FundWithNext;
 use std::sync::Arc;
