+++ src/cat-data-service/src/axum/v1/event/ballots.rs
@@ -1,8 +1,4 @@
-use crate::{
-    service::{handle_result, Error},
-    state::State,
-    types::SerdeType,
-};
+use crate::{axum::handle_result, service::Error, state::State, types::SerdeType};
 use axum::{extract::Path, routing::get, Router};
 use event_db::types::{ballot::ObjectiveBallots, event::EventId};
 use std::sync::Arc;
