+++ src/cat-data-service/src/axum/v1/jorm_mock/mod.rs
@@ -1,5 +1,6 @@
 use crate::{
-    service::{handle_result, Error},
+    axum::handle_result,
+    service::Error,
     state::State,
     types::jorm_mock::{AccountId, AccountVote, Fragments, FragmentsProcessingSummary},
 };
