+++ src/cat-data-service/src/service/service.rs
@@ -8,6 +8,8 @@ use crate::service::Error;
 use crate::service::api::mk_api;
 use crate::service::utilities::catch_panic::{set_panic_hook, ServicePanicHandler};
 use crate::service::utilities::metrics_tracing::{init_prometheus, log_requests};
+use crate::service::utilities::middleware::chain_axum::ChainAxum;
+
 use crate::settings::{get_api_hostnames, API_URL_PREFIX};
 
 use poem::endpoint::PrometheusExporter;
@@ -38,6 +40,7 @@ pub(crate) fn mk_app(hosts: Vec<String>, base_route: Option<Route>) -> impl Into
         )
         .with(Cors::new())
         .with(OpenTelemetryMetrics::new())
+        .with(ChainAxum::new()) // TODO: Remove this once all endpoints are ported.
         .with(CatchPanic::new().with_handler(ServicePanicHandler))
         .around(|ep, req| async move { Ok(log_requests(ep, req).await) })
 }
@@ -56,7 +59,7 @@ pub(crate) fn mk_app(hosts: Vec<String>, base_route: Option<Route>) -> impl Into
 /// * `Error::EventDbError` - cannot connect to the event db
 /// * `Error::IoError` - An IO error has occurred.
 ///
-pub async fn run_service(addr: &SocketAddr) -> Result<(), Error> {
+pub async fn run(addr: &SocketAddr) -> Result<(), Error> {
     tracing::info!("Starting Poem Service ...");
     tracing::info!("Listening on {addr}");
