+++ src/cat-data-service/src/service/utilities/middleware/mod.rs
@@ -1,0 +1,3 @@
+//! Custom POEM Middleware for this service.
+
+pub(crate) mod chain_axum;
