+++ src/cat-data-service/src/service/utilities/mod.rs
@@ -2,3 +2,4 @@
 //!
 pub(crate) mod catch_panic;
 pub(crate) mod metrics_tracing;
+pub(crate) mod middleware;
